Added all IE elemental metals to PeriodicTableElement, created MaterialMetalAluminum, MaterialMetalIron.
