+++ src/main/java/com/igteam/immersivegeology/common/blocks/EnumMetals.java
@@ -1,13 +1,14 @@
 package com.igteam.immersivegeology.common.blocks;
 
 import com.igteam.immersivegeology.api.materials.material_bases.MaterialMetalBase;
-import com.igteam.immersivegeology.common.materials.MaterialMetalCopper;
+import com.igteam.immersivegeology.common.materials.metals.*;
 
 public enum EnumMetals {
 
-    Copper(new MaterialMetalCopper())
-    /*, TODO add all Materials
+    Copper(new MaterialMetalCopper(), Type.IE_PURE),
     Aluminum(new MaterialMetalAluminum(), Type.IE_PURE),
+    Iron(new MaterialMetalIron(), Type.VANILLA)
+    /*, TODO add all Materials
     Lead(new MaterialMetalLead(), Type.IE_PURE),
     Silver(new MaterialMetalSilver(), Type.IE_PURE),
     Nickel(new MaterialMetalNickel(), Type.IE_PURE),
@@ -15,15 +16,24 @@ public enum EnumMetals {
     Constantan(new MaterialMetalConstantan(), Type.IE_ALLOY),
     Electrum(new MaterialMetalElectrum(), Type.IE_ALLOY),
     Steel(new MaterialMetalSteel(), Type.IE_ALLOY),
-    Iron(new MaterialMetalIron(), Type.VANILLA),
     Gold(new MaterialMetalGold(), Type.VANILLA) */
     ;
 
     public final MaterialMetalBase metal;
 
-    EnumMetals(MaterialMetalBase metal)
+    private final Type type;
+
+    EnumMetals(MaterialMetalBase metal, Type t)
     {
         this.metal = metal;
+        this.type = t;
     }
 
+    private enum Type {
+        VANILLA,
+        IE_PURE,
+        IE_ALLOY,
+        IG_PURE,
+        IG_ALLOY
+    }
 }
