+++ src/main/java/com/igteam/immersivegeology/common/materials/metals/MaterialMetalIron.java
@@ -1,0 +1,110 @@
+package com.igteam.immersivegeology.common.materials.metals;
+
+import blusunrize.immersiveengineering.ImmersiveEngineering;
+import com.igteam.immersivegeology.api.materials.PeriodicTableElement;
+import com.igteam.immersivegeology.api.materials.PeriodicTableElement.ElementProportion;
+import com.igteam.immersivegeology.api.materials.material_bases.MaterialMetalBase;
+import net.minecraft.item.Rarity;
+
+import javax.annotation.Nonnull;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+
+/**
+ * Created by JStocke12 on 27-03-2020.
+ */
+public class MaterialMetalIron extends MaterialMetalBase
+{
+	@Override
+	public String getName()
+	{
+		return "iron";
+	}
+
+	@Nonnull
+	@Override
+	public String getModID()
+	{
+		return "minecraft";
+	}
+
+	@Override
+	public Set<ElementProportion> getElements()
+	{
+		return new HashSet<>(Arrays.asList(
+				new ElementProportion(PeriodicTableElement.IRON)
+		));
+	}
+
+	@Override
+	public Rarity getRarity()
+	{
+		return Rarity.COMMON;
+	}
+
+	@Override
+	public int getBoilingPoint()
+	{
+		return 3153;
+	}
+
+	@Override
+	public int getMeltingPoint()
+	{
+		return 1811;
+	}
+
+	@Override
+	public int getColor(int temperature)
+	{
+		return 0xe39919;
+	}
+
+	//Needs to be changed in code for subtypes, such as sheetmetal
+	@Override
+	public float getHardness()
+	{
+		return 5.0F;
+	}
+
+	@Override
+	public float getMiningResistance()
+	{
+		return 10.0F;
+	}
+
+	@Override
+	public float getBlastResistance()
+	{
+		return 6;
+	}
+
+	//Copied from Immersive Intelligence (steel has i think 1.65, leaves 0.35)
+	@Override
+	public float getDensity()
+	{
+		return 1.25f;
+	}
+
+	//Stone pickaxe level
+	@Override
+	public int getBlockHarvestLevel()
+	{
+		return 1;
+	}
+
+	@Override
+	public EnumMetalType getMetalType()
+	{
+		return EnumMetalType.METAL;
+	}
+
+
+	/*@Nullable
+	@Override
+	public IItemTier getToolTier()
+	{
+		return IGContent.;
+	}*/
+}
